Update script.js
The only thing occurring in your code that concerns me is that I think I am indeed seeing a recursive method call where you're calling displayPattern() within the setStartPat() method yet setStartPat() in as a parameter of displayPattern(). It doesn't seem to cause improper behavior in the user experience, but in many situations it can be dangerous if not used with caution. I am curious though why displayPattern() is called at all in setStartPat()? If it is truly necessary and the recursive method call isn't a problem, then my comment would be to reorganize a little bit so that "setStartPat" only does what it says it does. Just some food for thought!

FYI, I didn't test any of my suggestions, so I apologize if there are any syntax errors, etc.